Políticas

+++ app/Http/Controllers/Admin/PostController.php
@@ -78,6 +78,8 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        $this->authorize('author', $post);
+
         $categories = Category::pluck('name', 'id');
         $tags = Tag::all();
 
@@ -124,6 +126,8 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('author', $post);
+
         $post->delete();
 
         return redirect()->route('admin.posts.index')->with('info', 'El post se eliminó con éxito');
